+++ packages/monaco-editor-comp/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to npm module [monaco-editor-comp](https://www.npmjs.com/package/monaco-editor-comp) are documented in this file.
 
+## [1.0.1]
+
+- Updated `monaco-editor-wrapper` to `1.1.1`
+
 ## [1.0.0]
 
 - Separated `monaco-editor-wrapper` from `monaco-editor-comp`